pyproject.toml
[project]
name = "pyrays"
version="0.0.0"
description = "TODO"
readme = "README.md"
requires-python = ">=3.7"
license = { file = "LICENSE.txt" }
authors = [
    { name = "Kian Shepherd", email = "kianshepherd73@gmail.com" },
]
keywords = ["pyrays"] # max of 5 keywords
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Other Environment",
    "Framework :: Flake8",
    "Framework :: Pytest",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = ["Pillow~=9.0"]

[project.optional-dependencies]
lint = [
    "flake8==4.0.1",
    "flake8-bugbear==22.1.11",
    "flake8-colors==0.1.9",
    "flake8-docstrings==1.6.0",
    "flake8-import-order==0.18.1",
    "pep8-naming==0.12.1",
    "pyproject-flake8==0.0.1a2",
]
test = [
    "coverage[toml]==6.3.2",
    "pytest==7.1.2",
    "pytest-cov==3.0.0",
    "imagehash~=4.0",
]
all = [
    "flake8==4.0.1",
    "flake8-bugbear==22.1.11",
    "flake8-colors==0.1.9",
    "flake8-docstrings==1.6.0",
    "flake8-import-order==0.18.1",
    "pep8-naming==0.12.1",
    "pyproject-flake8==0.0.1a2",
    "coverage[toml]==6.3.2",
    "pytest==7.1.2",
    "pytest-cov==3.0.0",
    "imagehash~=4.0",
]

[project.urls]
repository = "https://gitlab.com/kian_shepherd/pyrays"

[build-system]
requires = ["setuptools", "wheel", "setuptools-rust"]

[tool.flake8]
filename = "*.py"
exclude = [
    ".git",
    ".eggs",
    "*.egg",
    "*.egg-info",
    ".eggs",
    ".venv*",
    "venv*",
    "tests/*",
    "build/*",
]
show-source = "True"
max-complexity = 10
max-line-length = 100
per-file-ignores = [
    "__init__.py:F401,F403,D104", # Allow unused and * imports in __init__ files; this lets us expose definitions
    "__main__.py:F401,F403,D100" # Allow unused and * imports in __init__ files; this lets us expose definitions
]
ignore = [
    "D107", # missing docstring in __init__
    "E124", # closing bracket does not match visual indentation; Google's Python style guide is incompatible with this
    "I100", # import statements are in the wrong order
    "I202", # additional newline in a group of imports (We use three 3: built-in, third-party, local)
    "W503", # depracated warning - goes against PEP8
]

[tool.pytest.ini_options]
testpaths = [ 'tests' ]

[tool.coverage.run]
plugins = [ "Cython.Coverage" ]
concurrency = [ "multiprocessing" ]
parallel = true
branch = true
relative_files = true
data_file = "coverage/.coverage"
source = [ "pyrays" ]
omit = [
    "*/__main__.py",
    "*.so",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
precision = 2
exclude_lines = [
    "nocov",
]
